/home/roman/ext/OpenSCADA/Work/ChangeLog                                                                          2760/1025K               0% 2023-03-25 Roman (stable) - !!:LibsDB.vcaBase: wlb_Main.RootPgSo: The "Notification: type 2 (TextSpeech) (notify2)" method expanded by the RHVoice support         and appended by the builtin syntax highlight for BASH.     !!: The "Notification: type 1 (Buzzer) (notify1)" method rewrote for the long samples playback in the background,         that is with control the sample playing stop and restarting only in the case. Appended also by the syntax highlight. - !!!!:SYS,Transports,UI.Vision: The function's TSYS::strEncode() TSYS::ShieldSimb encoding renamed to TSYS::ShieldSymb,         moved to the function TSYS::strDecode() and in TSYS::strEncode() appended by the common one.     !!!!: And the UserAPI function SYS.strDecode() appended by the decoding method "ShieldSymb". - DAQ.{DAQGate,ModBus}: Status of the buffer of the asynchronous writing moved into the global context as "To write %d. ",         that is for the communication errors also, since the writing buffer is available here and costantly in the Asynchronous Writing.     Counters type switched to uint32_t for allowing the resetting and preventing the frozzing. - !:FIX:UI.{VCAEngine,Vision}: For encoding strings of the environment variables of the external notification scripts     there used the method TSYS::ShieldSymb and only in the symbol '"', for preventing encoding "'". - !!!:FIX:UI.VCAEngine: The dynamic translation of tr() fixed in the translation loss after the translation cache cleaning         and placing the spare messages to the project's table "Trs". - !:UI.Vision: Appended by overriding the Phonon notification method with a notification script declaration         in the prioritaty attribute "notifyVisVision{N}".     Enabled to notify the active alarms at the start, for the notification script as for Phonon. - !!:UI.WebUser: Expanded by a possibility of schedule the user procedures call in the service cycle,     for the background processing of some data.
git-svn-id: svn://oscada.org/trunk/OpenSCADA@2877 7bc60720-51fb-0310-bb35-cf765b5d350c

+++ src/moduls/daq/DAQGate/daq_gate.h
@@ -1,7 +1,7 @@
 
 //OpenSCADA module DAQ.DAQGate file: daq_gate.h
 /***************************************************************************
- *   Copyright (C) 2007-2022 by Roman Savochenko, <roman@oscada.org>       *
+ *   Copyright (C) 2007-2023 by Roman Savochenko, <roman@oscada.org>       *
  *                                                                         *
  *   This program is free software; you can redistribute it and/or modify  *
  *   it under the terms of the GNU General Public License as published by  *
@@ -169,7 +169,7 @@ class TMdContr: public TController
 
 	    ResMtx reqM, aWrRes, dataRes;
 
-	    float numR, numRA, numW, numRM, numWM;
+	    uint32_t numR, numRA, numW, numRM, numWM;
 	};
 	class SPrmsStack
 	{
